# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Restaurant Self-Service Kiosk API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /produtos:
    get:
      operationId: listarProdutos
      tags: [produtos]
      summary: List all products
      parameters:
        - name: categoria
          in: query
          required: false
          schema:
            type: integer
        - name: ativo
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"
    post:
      operationId: criarProduto
      tags: [produtos]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarProdutoInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
  /produtos/{id}:
    get:
      operationId: buscarProduto
      tags: [produtos]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
        "404":
          description: Product not found
    put:
      operationId: atualizarProduto
      tags: [produtos]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarProdutoInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
    delete:
      operationId: excluirProduto
      tags: [produtos]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
  /pedidos:
    get:
      operationId: listarPedidos
      tags: [pedidos]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [novo, em_preparo, pronto, entregue, cancelado]
        - name: data
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pedido"
    post:
      operationId: criarPedido
      tags: [pedidos]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarPedidoInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"
  /pedidos/{id}:
    get:
      operationId: buscarPedido
      tags: [pedidos]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"
  /pedidos/{id}/imprimir:
    post:
      operationId: imprimirPedido
      tags: [pedidos]
      summary: Re-print an order receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Print successful
...
  /pedidos/{id}/status:
    patch:
      operationId: atualizarStatusPedido
      tags: [pedidos]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"
  /relatorios/vendas:
    get:
      operationId: relatorioVendas
      tags: [pedidos]
      summary: Sales report
      parameters:
        - name: data
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatorioVendas"
...
components:
  schemas:
...
    Categoria:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        icone:
          type: string
          nullable: true
        imagemUrl:
          type: string
          nullable: true
        ordem:
          type: integer
      required:
        - id
        - nome
        - ordem
...
    Produto:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: string
          nullable: true
        preco:
          type: number
        categoriaId:
          type: integer
        imagem:
          type: string
          nullable: true
        ativo:
          type: boolean
        extras:
          type: array
          items:
            $ref: "#/components/schemas/Extra"
        categoria:
          $ref: "#/components/schemas/Categoria"
          nullable: true
      required:
        - id
        - nome
        - preco
        - categoriaId
        - ativo
        - extras
    Extra:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        preco:
          type: number
        tipo:
          type: string
          enum: [adicional, remocao, tamanho]
        grupoTitulo:
          type: string
        modoSelecao:
          type: string
          enum: [multipla, unica, sim_nao]
        maxSelecoes:
          type: integer
        obrigatorio:
          type: boolean
        ordemGrupo:
          type: integer
        ordemItem:
          type: integer
        imagem:
          type: string
          nullable: true
          description: URL da foto da opção no totem (/api/uploads/...)
      required:
        - id
        - nome
        - preco
        - tipo
        - grupoTitulo
        - modoSelecao
        - maxSelecoes
        - obrigatorio
        - ordemGrupo
        - ordemItem
    CriarProdutoInput:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
          nullable: true
        preco:
          type: number
        categoriaId:
          type: integer
        imagem:
          type: string
          nullable: true
        ativo:
          type: boolean
        extras:
          type: array
          items:
            $ref: "#/components/schemas/CriarExtraInput"
      required:
        - nome
        - preco
        - categoriaId
        - ativo
    CriarExtraInput:
      type: object
      properties:
        nome:
          type: string
        preco:
          type: number
        tipo:
          type: string
          enum: [adicional, remocao, tamanho]
        grupoTitulo:
          type: string
        modoSelecao:
          type: string
          enum: [multipla, unica, sim_nao]
        maxSelecoes:
          type: integer
        obrigatorio:
          type: boolean
        ordemGrupo:
          type: integer
        ordemItem:
          type: integer
        imagem:
          type: string
          nullable: true
      required:
        - nome
        - preco
        - tipo
    ItemPedido:
      type: object
      properties:
        id:
          type: integer
        pedidoId:
          type: integer
        produtoId:
          type: integer
        quantidade:
          type: integer
        preco:
          type: number
        observacoes:
          type: string
          nullable: true
        extrasIds:
          type: array
          items:
            type: integer
        produto:
          $ref: "#/components/schemas/Produto"
          nullable: true
      required:
        - id
        - pedidoId
        - produtoId
        - quantidade
        - preco
        - extrasIds
    CriarItemPedidoInput:
      type: object
      properties:
        produtoId:
          type: integer
        quantidade:
          type: integer
        observacoes:
          type: string
          nullable: true
        extrasIds:
          type: array
          items:
            type: integer
      required:
        - produtoId
        - quantidade
        - extrasIds
    Pedido:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: string
        status:
          type: string
          enum: [novo, em_preparo, pronto, entregue, cancelado]
        total:
          type: number
        tipoPedido:
          type: string
          enum: [comer_aqui, para_viagem]
        formaPagamento:
          type: string
          enum: [credito, debito, pix, nfc]
          nullable: true
        criadoEm:
          type: string
        itens:
          type: array
          items:
            $ref: "#/components/schemas/ItemPedido"
      required:
        - id
        - numero
        - status
        - total
        - tipoPedido
        - criadoEm
        - itens
    CriarPedidoInput:
      type: object
      properties:
        tipoPedido:
          type: string
          enum: [comer_aqui, para_viagem]
        formaPagamento:
          type: string
          enum: [credito, debito, pix, nfc]
        itens:
          type: array
          items:
            $ref: "#/components/schemas/CriarItemPedidoInput"
      required:
        - tipoPedido
        - formaPagamento
        - itens
    AtualizarStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [novo, em_preparo, pronto, entregue, cancelado]
      required:
        - status
    RelatorioVendas:
      type: object
      properties:
        totalVendas:
          type: number
        totalPedidos:
          type: integer
        ticketMedio:
          type: number
        produtosMaisVendidos:
          type: array
          items:
            $ref: "#/components/schemas/ProdutoVendido"
        vendasPorHora:
          type: array
          items:
            $ref: "#/components/schemas/VendaPorHora"
      required:
        - totalVendas
        - totalPedidos
        - ticketMedio
        - produtosMaisVendidos
        - vendasPorHora
    ProdutoVendido:
      type: object
      properties:
        produto:
          type: string
        quantidade:
          type: integer
        total:
          type: number
      required:
        - produto
        - quantidade
        - total
    VendaPorHora:
      type: object
      properties:
        hora:
          type: string
        total:
          type: number
        quantidade:
          type: integer
      required:
        - hora
        - total
        - quantidade
...